+++ clients/requests/auth.test.ts
@@ -1,8 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { createApp } from "vue";
-import { VueQueryPlugin, QueryClient } from "@tanstack/vue-query";
 import { useLogin, useLogout, useRegister } from "./auth";
 import { client } from "./client";
+import { withVueQuery } from "./test-utils";
 
 // Mock the client
 vi.mock("./client", () => ({
@@ -10,29 +9,6 @@ vi.mock("./client", () => ({
     POST: vi.fn(),
   },
 }));
-
-function withVueQuery(composable: () => unknown) {
-  let result;
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      mutations: {
-        retry: false,
-      },
-    },
-  });
-
-  const app = createApp({
-    setup() {
-      result = composable();
-      return () => {};
-    },
-  });
-
-  app.use(VueQueryPlugin, { queryClient });
-  app.mount(document.createElement("div"));
-
-  return [result, app] as const;
-}
 
 describe("auth requests", () => {
   const mockPOST = client.POST as ReturnType<typeof vi.fn>;
